Rename firestore log destination
