+++ lib/services/logger/destinations/firestore_log_destination.dart
